Refactor MCP host surface into explicit projection layers
